Script restructuring & cow UI
Broke up the Gamecontrol and movement scripts.

Removed the buying and selling buttons from the player UI.
Cows can now be selected, and bought and sold from there.

Also got rid of the dictionary, using just an int in the cow class to
link the cow to the gameobject now.