Last Changes - App Submitted!
This is the version going to market, version 0.6.

IF you want to update the app you must change the version number i.e
version (0.7, 0.8, 0.9) everytime you want to compile a new version to
upload to market. The keystore file must be kept included with the
project, (you can't update the app without it ever!)
There is also a password for the keystore file which I have, email if
you want the password.